Don't crash if we're missing headers
For a local, non-my-plex setup, Username and Transientoken can be blank.

Since empty inbound headers are being stripped, if the API client passes
blank values (which are correct, per above), the headers will be missing
and the plugin will throw an exception when attempting to access the
dictionary values.

Avoid this by setting the default value of all expected headers to ''.